just need to free allocations && images

+++ src/parcing.c
@@ -6,7 +6,7 @@
 /*   By: shrimech <shrimech@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/24 07:26:05 by shrimech          #+#    #+#             */
-/*   Updated: 2025/03/27 23:50:06 by shrimech         ###   ########.fr       */
+/*   Updated: 2025/03/29 05:36:26 by shrimech         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	elements_count(t_game *game)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	while (i < game->high)
@@ -40,8 +40,8 @@ int	elements_count(t_game *game)
 
 int	map_elements(t_game *game)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	while (game->map[i] != NULL)
